Many changes for alpha2 release
git-svn-id: http://svn.cs.brynmawr.edu/Myro@805 ed25d823-a23d-414a-8cb8-cff62a94a292

+++ branches/richard-dev-3/Services/Scribbler/ScribblerBase/ScribblerState.cs
@@ -433,7 +433,7 @@ namespace Myro.Services.Scribbler.ScribblerBase
         /// units: ms
         /// </summary>
         [DataMember]
-        public int Duration;
+        public double Duration;
 
         /// <summary>
         /// units: Hz
@@ -451,7 +451,7 @@ namespace Myro.Services.Scribbler.ScribblerBase
         {
         }
 
-        public PlayToneBody(int duration, int freq1, int freq2)
+        public PlayToneBody(double duration, int freq1, int freq2)
         {
             Duration = duration;
             Frequency1 = freq1;
